requirements.txt
openenv
torch
transformers
datasets
trl
unsloth
peft
accelerate
pandas
numpy
yfinance
ccxt
gymnasium
pettingzoo>=1.24.0
matplotlib
plotly
fastapi
uvicorn
langchain
langchain-openai
langchain-core
openai
python-dotenv
